Resolve textColor using view's trait collection
AvatarImageView renders text via a call to
NSString.draw(in:withAttributes:). If the text color is a dynamic color,
NSString.draw resolves the color based on the system user interface
style, so if overrideUserInterfaceStyle does not match the system
setting, the resolved color will not match the AvatarImageView's trait
collection. To fix this, this commit resolves the text color manually
based on the AvatarImageView's trait collection.